Extract util function
